[CIR-2309] Small refactors for improvements
