+++ app/uk/gov/hmrc/ninoinsightsstub/config/Module.scala
@@ -21,7 +21,7 @@ import com.google.inject.AbstractModule
 class Module extends AbstractModule {
 
   override def configure(): Unit = {
-
     bind(classOf[AppConfig]).asEagerSingleton()
+    bind(classOf[RejectList]).asEagerSingleton()
   }
 }
